foter format and media selector
