+++ FrontEnd/src/components/Footer/Footer.jsx
@@ -1,13 +1,17 @@
 import css from './Footer.module.css';
 import FooterPolicy from './FooterComponents/FooterPolicy';
 import FooterNavigation from './FooterComponents/FooterNavigation';
+import FooterAddress from './FooterComponents/FooterAddress';
+import useScrollToTop from '../../hooks/useScrollToTop';
 
 
 function Footer() {
+    useScrollToTop();
     return (
         <footer className={css['footer-content']}>
-                <FooterPolicy></FooterPolicy>
-                <FooterNavigation></FooterNavigation>
+            <FooterAddress></FooterAddress>
+            <FooterNavigation></FooterNavigation>
+            <FooterPolicy></FooterPolicy>
         </footer>
     );
 }
