+++ FrontEnd/src/components/Footer/FooterComponents/FooterNavigation.jsx
@@ -1,15 +1,9 @@
-import { useState, useEffect } from 'react';
 import { HashLink } from 'react-router-hash-link';
 import { Link } from 'react-router-dom';
-import axios from 'axios';
 import css from './FooterNavigation.module.css';
-import useScrollToTop from '../../../hooks/useScrollToTop';
+
 
 const PAGE_NAVIGATION_LINKS = [
-  {
-    title: 'Головна',
-    link: '/',
-  },
   {
     title: 'Компанії',
     link: '/profiles/companies',
@@ -17,10 +11,6 @@ const PAGE_NAVIGATION_LINKS = [
   {
     title: 'Стартапи',
     link: '/profiles/startups',
-  },
-  {
-    title: 'Про нас',
-    link: '/#about-us',
   },
 ];
 const SERVICES_LINKS = [
@@ -47,63 +37,33 @@ const SERVICES_LINKS = [
 ];
 
 function FooterNavigation() {
-  useScrollToTop();
-  const [contacts, setContacts] = useState(null);
-  const [error, setError] = useState(null);
 
-  useEffect(() => {
-    const fetchContacts = async () => {
-      try {
-        const url = `${process.env.REACT_APP_BASE_API_URL}/api/admin/contacts/`;
-        const response = await axios.get(url);
-        setContacts(response.data);
-      } catch (error) {
-        console.error('Error fetching contacts:', error);
-        setError('Не вдалося отримати контактну інформацію');
-      }
-    };
-
-    fetchContacts();
-  }, []);
 
   return (
     <div className={css['navigation-content']}>
-      <div className={css['navigation-content-section']}>
+      <div className={css['navigation-content__company']}>
+        <h3 className={css['navigation-content-company__header']}>Підприємства</h3>
         {PAGE_NAVIGATION_LINKS.map((element) => (
           <HashLink
             key={element.link}
-            className={css['navigation-content-section__text']}
+            className={css['navigation-content-company__text']}
             to={element.link}
           >
             {element.title}
           </HashLink>
         ))}
       </div>
-      <div className={css['navigation-content-section']}>
+      <div className={css['navigation-content__section']}>
+        <h3 className={css['navigation-content-section__header']}>Сектори</h3>
         {SERVICES_LINKS.map((element) => (
           <Link
-            className={css['navigation-content-section-service__text']}
+            className={css['navigation-content-section__text']}
             key={element.link}
             to={element.link}
           >
             {element.title}
           </Link>
         ))}
-      </div>
-      <div className={css['navigation-content-section']}>
-        <p className={css['navigation-content-section-service__text']}>Контакти</p>
-        {error ? (
-          <p className={css['navigation-content-section-service__text']} style={{ color: 'white' }}>
-            {error}
-          </p>
-        ) : contacts && (
-          <>
-            <p className={css['navigation-content-section-service__text']}>{contacts.email}</p>
-            <p className={css['navigation-content-section-service__text']}>{contacts.phone}</p>
-            <p className={css['navigation-content-section-service__text']}>{contacts.university}</p>
-            <p className={css['navigation-content-section-service__text']}>{contacts.address}</p>
-          </>
-        )}
       </div>
     </div>
   );
