+++ FrontEnd/src/components/Footer/FooterComponents/FooterNavigation.module.css
@@ -4,14 +4,28 @@
     gap: 140px;
 }
 
-.navigation-content-section {
+.navigation-content__company,
+.navigation-content__section {
     display: flex;
     flex-direction: column;
     align-items: flex-start;
     gap: 6px;
 }
 
-.navigation-content-section__text {
+.navigation-content-company__header,
+.navigation-content-section__header{
+    color: var(--footer-font-color);
+    font-feature-settings: 'calt' off;
+    text-decoration: none;
+    font-family: var(--font-main);
+    font-size: 14px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: 150%;
+    letter-spacing: -0.14px;
+}
+
+.navigation-content-company__text {
     color: var(--footer-font-color);
     font-feature-settings: 'calt' off;
     text-decoration: none;
@@ -23,7 +37,7 @@
     letter-spacing: -0.14px;
 }
 
-.navigation-content-section-service__text {
+.navigation-content-section__text {
     color: var(--footer-font-color);
     font-feature-settings: 'calt' off;
     text-decoration: none;
