+++ FrontEnd/src/components/Footer/FooterComponents/FooterPolicy.module.css
@@ -1,17 +1,24 @@
 .policy-content {
     display: flex;
     flex-direction: column;
-    align-items: flex-start;
-    gap: 79px;
+    padding-top: 1%;
 }
+
+.policy-divider {
+    width:auto;
+    height: 1px;
+    background:  #E2E5EB;
+}
+
 
 .policy-content-links {
     display: flex;
-    align-items: flex-start;
-    gap: 32px;
+    flex-direction: row;
+    gap: 12px;
 }
 
-.policy-content-links__text {
+.policy-content-links__text,
+.policy-content__copyright {
     color: var(--footer-links-font-color);
     text-align: center;
     font-feature-settings: 'calt' off;
@@ -25,7 +32,29 @@
     white-space: nowrap;
 }
 
-.policy-content__main-logo {
-    width: 199px;
-    height: 24px;
+@media (min-width: 375px) {
+    .policy-divider {
+        width:auto;
+        height: 1px;
+        background:  #E2E5EB;
+    }
+
+    .policy-content-links {
+        display: flex;
+        flex-direction: row;
+        gap: 12px;
+    }
+}
+
+@media (min-width: 768px) {
+    .policy-divider {
+        width:auto;
+        height: 1px;
+        background:  none;
+    }
+    .policy-content-links {
+        display: flex;
+        flex-direction: column;
+        gap: 12px;
+    }
 }
